Fix unit tests

+++ lib/emscripten/src/syscalls/mod.rs
@@ -405,6 +405,7 @@ pub fn ___syscall192(mut ctx: FunctionEnvMut<EmEnv>, _which: c_int, mut varargs:
 }
 
 /// lseek
+#[allow(clippy::unnecessary_cast)]
 pub fn ___syscall140(ctx: FunctionEnvMut<EmEnv>, _which: i32, mut varargs: VarArgs) -> i32 {
     // -> c_int
     debug!("emscripten::___syscall140 (lseek) {}", _which);
@@ -414,7 +415,7 @@ pub fn ___syscall140(ctx: FunctionEnvMut<EmEnv>, _which: i32, mut varargs: VarAr
     let result_ptr_value: WasmPtr<i64> = varargs.get(&ctx);
     let whence: i32 = varargs.get(&ctx);
     let offset = offset_low;
-    let ret = unsafe { lseek(fd, offset as _, whence) };
+    let ret = unsafe { lseek(fd, offset as _, whence) as i64 };
     let memory = ctx.data().memory(0);
     let memory = memory.view(&ctx);
